+++ channels_graphql_ws/graphql_ws_consumer.py
@@ -29,8 +29,7 @@ library `subscription-transport-ws` (which is used by Apollo).
 
 NOTE: Links based on which this functionality is implemented:
 - Protocol description:
-  https://github.com/apollographql/subscriptions-transport-ws/blob/master/PROTOCOL.md
-  https://github.com/apollographql/subscriptions-transport-ws/blob/master/src/message-types.ts
+  https://github.com/enisdenjo/graphql-ws/blob/master/PROTOCOL.md
 - ASGI specification for WebSockets:
   https://github.com/django/asgiref/blob/master/specs/www.rst#websocket
 - GitHubGist with the root of inspiration:
@@ -77,7 +76,7 @@ from .serializer import Serializer
 LOG = logging.getLogger(__name__)
 
 # WebSocket subprotocol used for the GraphQL.
-GRAPHQL_WS_SUBPROTOCOL = "graphql-ws"
+GRAPHQL_WS_SUBPROTOCOL = "graphql-transport-ws"
 
 
 class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
@@ -87,8 +86,8 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
     connection to a single client.
 
     This class implements the WebSocket-based GraphQL protocol used by
-    `subscriptions-transport-ws` library (used by Apollo):
-    https://github.com/apollographql/subscriptions-transport-ws/blob/master/PROTOCOL.md
+    `graphql-ws` library (used by Apollo):
+    https://github.com/enisdenjo/graphql-ws/blob/master/PROTOCOL.md
     """
 
     # ----------------------------------------------------------------- PUBLIC INTERFACE
@@ -96,9 +95,6 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
     # Overwrite this in the subclass to specify the GraphQL schema which
     # processes GraphQL queries.
     schema: graphene.Schema
-
-    # The interval to send keepalive messages to the clients (seconds).
-    send_keepalive_every: Optional[float] = None
 
     # Set to `True` to process requests (i.e. GraphQL documents) from
     # a client in order of arrival, which is the same as sending order,
@@ -110,7 +106,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
     # are still processed asynchronously. Useful for tests.
     strict_ordering: bool = False
 
-    # When set to `True` the server will send an empty data message in
+    # When set to `True` the server will send an empty next message in
     # response to the subscription. This is needed to let client know
     # when the subscription activates, so he can be sure he doesn't miss
     # any notifications. Disabled by default, cause this is an extension
@@ -220,7 +216,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
 
         # Crafty weak collection with per-operation locks. It holds a
         # mapping from the operaion id (protocol message id) to the
-        # `asyncio.Lock` used to serialize processing of start & stop
+        # `asyncio.Lock` used to serialize processing of subscribe & complete
         # requests. Since the collection is weak, it automatically
         # throws away items when locks are garbage collected.
         self._operation_locks: weakref.WeakValueDictionary = (
@@ -318,7 +314,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
     async def receive_json(self, content):  # pylint: disable=arguments-differ
         """Process WebSocket message received from the client.
 
-        NOTE: We force 'STOP' message processing to wait until 'START'
+        NOTE: We force 'COMPLETE' message processing to wait until 'SUBSCRIBE'
         with the same operation id finishes (if it is running). This
         protects us from race conditions which may happen when a client
         stops operation immediately after starting it. An illustrative
@@ -331,17 +327,14 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         msg_type = content["type"].upper()
 
         if msg_type == "CONNECTION_INIT":
-            task = self._on_gql_connection_init(payload=content["payload"])
+            task = self._on_gql_connection_init(payload=content.get("payload"))
 
-        elif msg_type == "CONNECTION_TERMINATE":
-            task = self._on_gql_connection_terminate()
-
-        elif msg_type == "START":
+        elif msg_type == "SUBSCRIBE":
             op_id = content["id"]
 
             # Create and lock a mutex for this particular operation id,
-            # so STOP processing for the same operation id will wait
-            # until START processing finishes. Locks are stored in a
+            # so COMPLETE processing for the same operation id will wait
+            # until SUBSCRIBE processing finishes. Locks are stored in a
             # weak collection so we do not have to manually clean it up.
             if op_id in self._operation_locks:
                 raise graphql.error.GraphQLError(
@@ -351,28 +344,34 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
             self._operation_locks[op_id] = op_lock
             await op_lock.acquire()
 
-            async def on_start():
+            async def on_subscribe():
                 try:
                     # User hook which raises to cancel processing.
                     await self.on_operation(op_id, payload=content["payload"])
-                    # START message processing.
-                    await self._on_gql_start(op_id, payload=content["payload"])
+                    # SUBSCRIBE message processing.
+                    await self._on_gql_subscribe(op_id, payload=content["payload"])
                 except Exception as ex:  # pylint: disable=broad-except
                     await self._send_gql_error(op_id, ex)
                 finally:
                     op_lock.release()
 
-            task = on_start()
+            task = on_subscribe()
 
-        elif msg_type == "STOP":
+        elif msg_type == "COMPLETE":
             op_id = content["id"]
 
-            async def on_stop():
-                # Wait until START message processing finishes, if any.
+            async def on_complete():
+                # Wait until SUBSCRIBE message with same id processing finishes, if any.
                 async with self._operation_locks.setdefault(op_id, asyncio.Lock()):
-                    await self._on_gql_stop(op_id)
+                    await self._on_gql_complete(op_id)
 
-            task = on_stop()
+            task = on_complete()
+
+        elif msg_type == "PING":
+            task = self._on_gql_ping()
+
+        elif msg_type == "PONG":
+            task = self._on_gql_pong()
 
         else:
             task = self._send_gql_error(
@@ -449,12 +448,14 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
 
         # Send messages which look like user unsubscribes from all
         # subscriptions in the subscription group. This saves us from
-        # thinking about raise condition between subscription and
+        # thinking about race condition between subscription and
         # unsubscription.
         if self._sids_by_group[group]:
             await asyncio.wait(
                 [
-                    asyncio.create_task(self.receive_json({"type": "stop", "id": sid}))
+                    asyncio.create_task(
+                        self.receive_json({"type": "complete", "id": sid})
+                    )
                     for sid in self._sids_by_group[group]
                 ]
             )
@@ -463,9 +464,6 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
 
     async def _on_gql_connection_init(self, payload):
         """Process the CONNECTION_INIT message.
-
-        Start sending keepalive messages if `send_keepalive_every` set.
-        Respond with either CONNECTION_ACK or CONNECTION_ERROR message.
 
         NOTE: Depending on the value of the `strict_ordering` setting
         this method is either awaited directly or offloaded to an async
@@ -488,35 +486,23 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         else:
             # Send CONNECTION_ACK message.
             await self._send_gql_connection_ack()
-            # If keepalive enabled then send one message immediately and
-            # schedule periodic messages.
-            if self.send_keepalive_every is not None:
-                send_keepalive_every = self.send_keepalive_every
 
-                async def keepalive_sender():
-                    """Send keepalive messages periodically."""
-                    while True:
-                        await asyncio.sleep(send_keepalive_every)
-                        await self._send_gql_connection_keep_alive()
+    async def _on_gql_ping(self):
+        """Process the PING message.
 
-                self._keepalive_task = asyncio.create_task(keepalive_sender())
-                # Immediately send keepalive message cause it is
-                # required by the protocol description.
-                await self._send_gql_connection_keep_alive()
-
-    async def _on_gql_connection_terminate(self):
-        """Process the CONNECTION_TERMINATE message.
-
-        NOTE: Depending on the value of the `strict_ordering` setting
-        this method is either awaited directly or offloaded to an async
-        task. See the `receive_json` handler.
+        Send PONG message as response.
         """
+        await self._send_gql_pong()
 
-        # Close the connection.
-        await self.close(code=1000)
+    async def _on_gql_pong(self):
+        """Process the PONG message.
 
-    async def _on_gql_start(self, op_id, payload):
-        """Process the START message.
+        Send PING message as response.
+        """
+        await self._send_gql_ping()
+
+    async def _on_gql_subscribe(self, op_id, payload):
+        """Process the SUBSCRIBE message.
 
         Handle the message with query, mutation or subscription request.
 
@@ -546,7 +532,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
                 op_name, query
             )
             if errors:
-                await self._send_gql_data(op_id, None, errors)
+                await self._send_gql_next(op_id, None, errors)
                 await self._send_gql_complete(op_id)
                 return
             # Assert values are not None to suppress MyPy complains.
@@ -563,14 +549,14 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
 
                 # This returns asynchronous generator or ExecutionResult
                 # instance in case of error.
-                subscr_result = await self._on_gql_start__subscribe(
+                subscr_result = await self._on_gql_subscribe__create_subscription(
                     doc_ast,
                     operation_name=op_name,
                     root_value=None,
                     variable_values=variables,
                     context_value=context,
                     subscribe_field_resolver=functools.partial(
-                        self._on_gql_start__initialize_subscription_stream,
+                        self._on_gql_subscribe__initialize_subscription_stream,
                         op_id,
                         op_name,
                     ),
@@ -586,7 +572,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
                     # stream to guarantee that no notifications are sent
                     # before the subscription confirmation message.
                     if self.confirm_subscriptions:
-                        await self._send_gql_data(
+                        await self._send_gql_next(
                             op_id,
                             data=self.subscription_confirmation_message["data"],
                             errors=self.subscription_confirmation_message["errors"],
@@ -603,7 +589,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
                                 # when we have something to send.
                                 if item.data or item.errors:
                                     try:
-                                        await self._send_gql_data(
+                                        await self._send_gql_next(
                                             op_id, item.data, item.errors
                                         )
                                     except asyncio.CancelledError:
@@ -616,7 +602,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
                                 op_id,
                                 exc_info=ex,
                             )
-                            await self._send_gql_data(op_id, None, [ex])
+                            await self._send_gql_next(op_id, None, [ex])
 
                     # We need to end this task when client drops
                     # connection or unsubscribes, so lets store it.
@@ -690,7 +676,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
                             variables,
                         )
             # Respond to a query or mutation immediately.
-            await self._send_gql_data(
+            await self._send_gql_next(
                 op_id, operation_result.data, operation_result.errors
             )
             await self._send_gql_complete(op_id)
@@ -701,7 +687,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
                 LOG.warning(
                     "GraphQL error! Operation %s(%s).", op_name, op_id, exc_info=True
                 )
-                await self._send_gql_data(op_id, None, [ex])
+                await self._send_gql_next(op_id, None, [ex])
                 await self._send_gql_complete(op_id)
             else:
                 # Respond with general error responce.
@@ -775,7 +761,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
 
         return doc_ast, op_ast, None
 
-    async def _on_gql_start__subscribe(
+    async def _on_gql_subscribe__create_subscription(
         self,
         document: graphql.DocumentNode,
         root_value: Any = None,
@@ -793,8 +779,8 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         the GraphQL-core library v3.2.3 improved to support middlewares
         and user defined execution_context_class.
 
-        This is a part of START message processing routine so the name
-        prefixed with `_on_gql_start__` to make this explicit.
+        This is a part of SUBSCRIBE message processing routine so the name
+        prefixed with `_on_gql_subscribe__` to make this explicit.
         """
 
         result_or_stream = await graphql.create_source_event_stream(
@@ -844,7 +830,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         # Map every source value to a ExecutionResult value.
         return graphql.MapAsyncIterator(result_or_stream, map_source_to_response)
 
-    async def _on_gql_start__initialize_subscription_stream(
+    async def _on_gql_subscribe__initialize_subscription_stream(
         self,
         operation_id: int,
         operation_name: str,
@@ -855,11 +841,11 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
     ):
         """Create asynchronous generator with subscription events.
 
-        Called inside `_on_gql_start__subscribe` function by
+        Called inside `_on_gql_subscribe__create_subscription` function by
         graphql-core as `subscribe_field_resolver` argument.
 
-        This is a part of START message processing routine so the name
-        prefixed with `_on_gql_start__` to make this explicit.
+        This is a part of SUBSCRIBE message processing routine so the name
+        prefixed with `_on_gql_subscribe__` to make this explicit.
         """
         # Graphene stores original subscription class in `graphene_type`
         # field of `return_type` object. Since subscriptions are build
@@ -999,7 +985,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         )
 
         # For each notification (event) yielded from this function the
-        # `_on_gql_start__subscribe` function will call subscription
+        # `_on_gql_subscribe__create_subscription` function will call subscription
         # resolver (`publish`) via `graphql.execute` method.
         while True:
             with notification_queue_lock:
@@ -1009,8 +995,8 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
             with notification_queue_lock:
                 notification_queue.task_done()
 
-    async def _on_gql_stop(self, op_id):
-        """Process the STOP message.
+    async def _on_gql_complete(self, op_id):
+        """Process the COMPLETE message.
 
         Handle an unsubscribe request.
 
@@ -1020,8 +1006,8 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         """
         LOG.debug("Stop handling or unsubscribe operation %s.", op_id)
 
-        # Currently only subscriptions can be stopped. But we see but
-        # some clients (e.g. GraphiQL) send the stop message even for
+        # Currently only subscriptions can be stopped. But we see that
+        # some clients (e.g. GraphiQL) send the complete message even for
         # queries and mutations. We also see that the Apollo server
         # ignores such messages, so we ignore them as well.
         if op_id not in self._subscriptions:
@@ -1071,6 +1057,14 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         """Sent in reply to the `connection_init` request."""
         await self.send_json({"type": "connection_ack"})
 
+    async def _send_gql_ping(self):
+        """Sent in reply to the `pong` request."""
+        await self.send_json({"type": "ping"})
+
+    async def _send_gql_pong(self):
+        """Sent in reply to the `ping` request."""
+        await self.send_json({"type": "pong"})
+
     async def _send_gql_connection_error(self, error: Exception):
         """Connection error sent in reply to the `connection_init`."""
         LOG.warning("GraphQL connection error: %s!", error, exc_info=error)
@@ -1078,10 +1072,10 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
             {"type": "connection_error", "payload": self._format_error(error)}
         )
 
-    async def _send_gql_data(
+    async def _send_gql_next(
         self, op_id, data: Optional[dict], errors: Optional[Iterable[Exception]]
     ):
-        """Send GraphQL `data` message to the client.
+        """Send GraphQL `next` message with data to the client.
 
         Args:
             data: Dict with GraphQL query response.
@@ -1102,7 +1096,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
 
         await self.send_json(
             {
-                "type": "data",
+                "type": "next",
                 "id": op_id,
                 "payload": {
                     "data": data,
@@ -1126,7 +1120,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         It can be an unexpected or unexplained GraphQL execution error
         or a bug in the code. It is unlikely that this is GraphQL
         validation errors (such errors are part of data message and
-        must be sent by the `_send_gql_data` method).
+        must be sent by the `_send_gql_next` method).
 
         Args:
             op_id: Id of the operation that failed on the server.
@@ -1152,10 +1146,6 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         """
         await self.send_json({"type": "complete", "id": op_id})
 
-    async def _send_gql_connection_keep_alive(self):
-        """Send the keepalive (ping) message."""
-        await self.send_json({"type": "ka"})
-
     # ---------------------------------------------------------------------- AUXILIARIES
 
     @staticmethod
@@ -1170,7 +1160,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         for a client like this:
             {
                 "id": "NNN",
-                "type": "data",
+                "type": "next",
                 "payload": {
                     "data": {...},
                     "errors": [{