+++ tests/test_serialization.py
@@ -125,7 +125,7 @@ async def test_models_serialization_with_nested_db_query(gql, transactional_db):
     print("Subscribe to receive subscription notifications.")
 
     sub_id = await client.send(
-        msg_type="start",
+        msg_type="subscribe",
         payload={
             "query": textwrap.dedent(
                 """
@@ -234,7 +234,7 @@ async def test_models_serialization(gql, transactional_db):
     print("Subscribe to receive subscription notifications.")
 
     sub_id = await client.send(
-        msg_type="start",
+        msg_type="subscribe",
         payload={
             "query": textwrap.dedent(
                 """
@@ -351,7 +351,7 @@ async def test_timestamps_serialization(gql, transactional_db):
     print("Subscribe to receive subscription notifications.")
 
     sub_id = await client.send(
-        msg_type="start",
+        msg_type="subscribe",
         payload={
             "query": textwrap.dedent(
                 """